FIX SPEC #7013 : use database type "numeric" to store monetary values

+++ htdocs/install/mysql/tables/llx_accounting_bookkeeping_tmp.sql
@@ -32,11 +32,11 @@ CREATE TABLE llx_accounting_bookkeeping_tmp
   numero_compte         varchar(32),		        -- FEC:CompteNum	| account number
   label_compte          varchar(255) NOT NULL,		-- FEC:CompteLib	| label of account
   label_operation       varchar(255),				-- FEC:EcritureLib	| label of the operation
-  debit                 double(24,8) NOT NULL,		-- FEC:Debit
-  credit                double(24,8) NOT NULL,		-- FEC:Credit
-  montant               double(24,8) NOT NULL,		-- FEC:Montant (Not necessary)
+  debit                 numeric(24,8) NOT NULL,		-- FEC:Debit
+  credit                numeric(24,8) NOT NULL,		-- FEC:Credit
+  montant               numeric(24,8) NOT NULL,		-- FEC:Montant (Not necessary)
   sens                  varchar(1) DEFAULT NULL,	-- FEC:Sens (Not necessary)
-  multicurrency_amount  double(24,8),				-- FEC:Montantdevise
+  multicurrency_amount  numeric(24,8),				-- FEC:Montantdevise
   multicurrency_code    varchar(255),				-- FEC:Idevise
   lettering_code        varchar(255),				-- FEC:EcritureLet
   date_lettering        datetime,					-- FEC:DateLet
