+++ htdocs/install/mysql/tables/llx_bordereau_cheque.sql
@@ -25,10 +25,10 @@ create table llx_bordereau_cheque
 (
   rowid             integer AUTO_INCREMENT PRIMARY KEY,
   ref               varchar(30) NOT NULL,					-- ref
-  ref_ext	          varchar(255),							-- ref_ext
+  ref_ext           varchar(255),							-- ref_ext
   datec             datetime NOT NULL,
   date_bordereau    date,
-  amount            double(24,8) NOT NULL,
+  amount            numeric(24,8) NOT NULL,
   nbcheque          smallint NOT NULL,
   fk_bank_account   integer,
   fk_user_author    integer,
